@@ -8,15 +8,15 @@ android {
     
     // Note to Self:
     // Update .travis.yml to ensure build stability whenever platform version is upgraded
-    compileSdkVersion 30
+    compileSdkVersion 31
     buildToolsVersion = '29.0.3'
 
     defaultConfig {
         applicationId "com.ae.apps.stickerapp"
         minSdkVersion 18
-        targetSdkVersion 30
+        targetSdkVersion 31
         versionCode 18
-        versionName "1.7.1"
+        versionName "1.8.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         def contentProviderAuthority = applicationId + ".stickercontentprovider"
         // Creates a placeholder property to use in the manifest.
