@@ -8,59 +8,58 @@
 
 package com.ae.apps.stickerapp;
 
-import android.content.Context;
 import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.ae.apps.stickerapp.analytics.Analytics;
+import com.facebook.drawee.backends.pipeline.Fresco;
+import com.facebook.drawee.interfaces.DraweeController;
 import com.facebook.drawee.view.SimpleDraweeView;
-import com.google.android.gms.ads.AdRequest;
-import com.google.android.gms.ads.AdView;
-import com.google.android.gms.ads.MobileAds;
 
 public class StickerPreviewAdapter extends RecyclerView.Adapter<StickerPreviewViewHolder> {
 
+    private static final float COLLAPSED_STICKER_PREVIEW_BACKGROUND_ALPHA = 1f;
+    private static final float EXPANDED_STICKER_PREVIEW_BACKGROUND_ALPHA = 0.2f;
+
     @NonNull
-    private StickerPack stickerPack;
-    @NonNull
-    final Context context;
+    private final StickerPack stickerPack;
+
     private final int cellSize;
-    private int cellLimit;
-    private int cellPadding;
+    private final int cellLimit;
+    private final int cellPadding;
     private final int errorResource;
-    private Analytics analytics;
+    private final SimpleDraweeView expandedStickerPreview;
+
     private final LayoutInflater layoutInflater;
+    private RecyclerView recyclerView;
+    private View clickedStickerPreview;
+    float expandedViewLeftX;
+    float expandedViewTopY;
 
     StickerPreviewAdapter(
-            @NonNull final Context context,
             @NonNull final LayoutInflater layoutInflater,
             final int errorResource,
             final int cellSize,
             final int cellPadding,
-            @NonNull final StickerPack stickerPack) {
-        this.context = context;
+            @NonNull final StickerPack stickerPack,
+            final SimpleDraweeView expandedStickerView) {
         this.cellSize = cellSize;
         this.cellPadding = cellPadding;
         this.cellLimit = 0;
         this.layoutInflater = layoutInflater;
         this.errorResource = errorResource;
         this.stickerPack = stickerPack;
-
-        MobileAds.initialize(context, initializationStatus -> {
-        });
-        analytics = Analytics.getInstance(context);
+        this.expandedStickerPreview = expandedStickerView;
     }
 
     @NonNull
     @Override
     public StickerPreviewViewHolder onCreateViewHolder(@NonNull final ViewGroup viewGroup, final int i) {
-        View itemView = layoutInflater.inflate(R.layout.sticker_image, viewGroup, false);
+        View itemView = layoutInflater.inflate(R.layout.sticker_image_item, viewGroup, false);
         StickerPreviewViewHolder vh = new StickerPreviewViewHolder(itemView);
 
         ViewGroup.LayoutParams layoutParams = vh.stickerPreviewView.getLayoutParams();
@@ -75,35 +74,126 @@ public class StickerPreviewAdapter extends RecyclerView.Adapter<StickerPreviewVi
     @Override
     public void onBindViewHolder(@NonNull final StickerPreviewViewHolder stickerPreviewViewHolder, final int i) {
         stickerPreviewViewHolder.stickerPreviewView.setImageResource(errorResource);
-        final String stickerFileName = stickerPack.getStickers().get(i).imageFileName;
-        final Uri stickerAssetUri = StickerPackLoader.getStickerAssetUri(stickerPack.identifier, stickerFileName);
-        stickerPreviewViewHolder.stickerPreviewView.setImageURI(stickerAssetUri);
-        stickerPreviewViewHolder.stickerPreviewView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                analytics.logEvent(Analytics.VIEW_ITEM, "icon_preview", stickerPack.name);
+        stickerPreviewViewHolder.stickerPreviewView.setImageURI(StickerPackLoader.getStickerAssetUri(stickerPack.identifier, stickerPack.getStickers().get(i).imageFileName));
+        stickerPreviewViewHolder.stickerPreviewView.setOnClickListener(v -> expandPreview(i, stickerPreviewViewHolder.stickerPreviewView));
+    }
 
-                View dialogView = LayoutInflater.from(context).inflate(R.layout.sticker_dialog, null, false);
-                SimpleDraweeView simpleDraweeView = dialogView.findViewById(R.id.sticker_preview_in_dialog);
-                AdView mAdView = dialogView.findViewById(R.id.adView);
-                View closeButton = dialogView.findViewById(R.id.btnClose);
-                simpleDraweeView.setImageURI(stickerAssetUri);
+    @Override
+    public void onAttachedToRecyclerView(@NonNull RecyclerView recyclerView) {
+        super.onAttachedToRecyclerView(recyclerView);
+        this.recyclerView = recyclerView;
+        recyclerView.addOnScrollListener(hideExpandedViewScrollListener);
+    }
 
-                AlertDialog.Builder builder = new AlertDialog.Builder(context);
-                builder.setView(dialogView);
-                AlertDialog alertDialog = builder.create();
-                alertDialog.show();
+    @Override
+    public void onDetachedFromRecyclerView(@NonNull RecyclerView recyclerView) {
+        super.onDetachedFromRecyclerView(recyclerView);
+        recyclerView.removeOnScrollListener(hideExpandedViewScrollListener);
+        this.recyclerView = null;
+    }
 
-                closeButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        alertDialog.dismiss();
+    private final RecyclerView.OnScrollListener hideExpandedViewScrollListener =
+            new RecyclerView.OnScrollListener() {
+                @Override
+                public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
+                    super.onScrolled(recyclerView, dx, dy);
+                    if (dx != 0 || dy != 0) {
+                        hideExpandedStickerPreview();
                     }
-                });
+                }
+            };
 
-                mAdView.loadAd(new AdRequest.Builder().build());
+    private void positionExpandedStickerPreview(int selectedPosition) {
+        if (expandedStickerPreview != null) {
+            // Calculate the view's center (x, y), then use expandedStickerPreview's height and
+            // width to
+            // figure out what where to position it.
+            final ViewGroup.MarginLayoutParams recyclerViewLayoutParams =
+                    ((ViewGroup.MarginLayoutParams) recyclerView.getLayoutParams());
+            final int recyclerViewLeftMargin = recyclerViewLayoutParams.leftMargin;
+            final int recyclerViewRightMargin = recyclerViewLayoutParams.rightMargin;
+            final int recyclerViewWidth = recyclerView.getWidth();
+            final int recyclerViewHeight = recyclerView.getHeight();
+
+            final StickerPreviewViewHolder clickedViewHolder =
+                    (StickerPreviewViewHolder)
+                            recyclerView.findViewHolderForAdapterPosition(selectedPosition);
+            if (clickedViewHolder == null) {
+                hideExpandedStickerPreview();
+                return;
             }
-        });
+            clickedStickerPreview = clickedViewHolder.itemView;
+            final float clickedViewCenterX =
+                    clickedStickerPreview.getX()
+                            + recyclerViewLeftMargin
+                            + clickedStickerPreview.getWidth() / 2f;
+            final float clickedViewCenterY =
+                    clickedStickerPreview.getY() + clickedStickerPreview.getHeight() / 2f;
+
+            expandedViewLeftX = clickedViewCenterX - expandedStickerPreview.getWidth() / 2f;
+            expandedViewTopY = clickedViewCenterY - expandedStickerPreview.getHeight() / 2f;
+
+            // If the new x or y positions are negative, anchor them to 0 to avoid clipping
+            // the left side of the device and the top of the recycler view.
+            expandedViewLeftX = Math.max(expandedViewLeftX, 0);
+            expandedViewTopY = Math.max(expandedViewTopY, 0);
+
+            // If the bottom or right sides are clipped, we need to move the top left positions
+            // so that those sides are no longer clipped.
+            final float adjustmentX =
+                    Math.max(
+                            expandedViewLeftX
+                                    + expandedStickerPreview.getWidth()
+                                    - recyclerViewWidth
+                                    - recyclerViewRightMargin,
+                            0);
+            final float adjustmentY =
+                    Math.max(expandedViewTopY + expandedStickerPreview.getHeight() - recyclerViewHeight, 0);
+
+            expandedViewLeftX -= adjustmentX;
+            expandedViewTopY -= adjustmentY;
+
+
+            expandedStickerPreview.setX(expandedViewLeftX);
+            expandedStickerPreview.setY(expandedViewTopY);
+        }
+    }
+
+    private void expandPreview(int position, View clickedStickerPreview) {
+        if (isStickerPreviewExpanded()) {
+            hideExpandedStickerPreview();
+            return;
+        }
+
+        this.clickedStickerPreview = clickedStickerPreview;
+        if (expandedStickerPreview != null) {
+            positionExpandedStickerPreview(position);
+
+            final Uri stickerAssetUri = StickerPackLoader.getStickerAssetUri(stickerPack.identifier, stickerPack.getStickers().get(position).imageFileName);
+            DraweeController controller = Fresco.newDraweeControllerBuilder()
+                    .setUri(stickerAssetUri)
+                    .setAutoPlayAnimations(true)
+                    .build();
+            expandedStickerPreview.setImageResource(errorResource);
+            expandedStickerPreview.setController(controller);
+
+            expandedStickerPreview.setVisibility(View.VISIBLE);
+            recyclerView.setAlpha(EXPANDED_STICKER_PREVIEW_BACKGROUND_ALPHA);
+
+            expandedStickerPreview.setOnClickListener(v -> hideExpandedStickerPreview());
+        }
+    }
+
+    public void hideExpandedStickerPreview() {
+        if (isStickerPreviewExpanded() && expandedStickerPreview != null) {
+            clickedStickerPreview.setVisibility(View.VISIBLE);
+            expandedStickerPreview.setVisibility(View.INVISIBLE);
+            recyclerView.setAlpha(COLLAPSED_STICKER_PREVIEW_BACKGROUND_ALPHA);
+        }
+    }
+
+    private boolean isStickerPreviewExpanded() {
+        return expandedStickerPreview != null && expandedStickerPreview.getVisibility() == View.VISIBLE;
     }
 
     @Override
